moving around setups
